# api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenRHP API",
    "version": "0.1.0-dev",
    "description": "OpenReverseHomeProxy. Authenticated API shared by the UI, CLI and agents. Development status is exposed by capabilities; unavailable network features fail preflight."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "operationId": "status",
        "summary": "Current decision, source health and runtime status",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        }
      }
    },
    "/capabilities": {
      "get": {
        "operationId": "capabilities",
        "summary": "Detected platform capabilities and explicit limitations",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        }
      }
    },
    "/preflight": {
      "get": {
        "operationId": "preflight",
        "summary": "Read-only compatibility and network readiness report",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        }
      }
    },
    "/openapi": {
      "get": {
        "operationId": "openapi",
        "summary": "Fetch the public OpenAPI contract",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        }
      }
    },
    "/config": {
      "get": {
        "operationId": "getConfig",
        "summary": "Read configuration with source secrets removed",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        }
      },
      "put": {
        "operationId": "replaceConfig",
        "summary": "Replace configuration; all source settings must be supplied",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            },
            "description": "Unique for this logical mutation. Reuse only with identical method/path/body and If-Match. Retained for 24h; interrupted work is reported, never blindly rerun."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Quoted current config revision, e.g. \"1\"."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config"
              }
            }
          }
        }
      }
    },
    "/config/validate": {
      "post": {
        "operationId": "validateConfig",
        "summary": "Validate without changing configuration or routes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient scope or origin not allowed"
          },
          "409": {
            "description": "Revision, pending transaction or idempotency conflict"
          },
          "422": {
            "description": "Unsupported or invalid configuration"
          },
          "429": {
            "description": "Rate or queue limit"
          },
          "503": {
            "description": "Capability unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Config": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "revision",
          "role",
          "sources",
          "targets",
          "policy",
          "probes",
          "network"
        ],
        "properties": {
          "schema_version": {
            "const": 1
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          },
          "role": {
            "enum": [
              "gateway",
              "node"
            ]
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Source"
            }
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Target"
            }
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "probes": {
            "$ref": "#/components/schemas/ProbeSettings"
          },
          "network": {
            "$ref": "#/components/schemas/Network"
          }
        }
      },
...
    }
  }
}